+++ client/src/components/Header.js
@@ -21,66 +21,7 @@ import ClickAwayListener from "@material-ui/core/ClickAwayListener";
 import CloudUploadIcon from "@material-ui/icons/CloudUpload";
 import DashboardIcon from "@material-ui/icons/Dashboard"
 import UpdateIcon from "@material-ui/icons/Update"
-
-const drawerWidth = 240;
-
-const styles = theme => ({
-  root: {
-    display: "flex",
-    marginBottom: theme.spacing.unit * 15
-  },
-  appBar: {
-    transition: theme.transitions.create(["margin", "width"], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen
-    })
-  },
-  appBarShift: {
-    width: `calc(100% - ${drawerWidth}px)`,
-    marginLeft: drawerWidth,
-    transition: theme.transitions.create(["margin", "width"], {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen
-    })
-  },
-  menuButton: {
-    marginLeft: 12,
-    marginRight: 20
-  },
-  hide: {
-    display: "none"
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0
-  },
-  drawerPaper: {
-    width: drawerWidth
-  },
-  drawerHeader: {
-    display: "flex",
-    alignItems: "center",
-    padding: "0 8px",
-    ...theme.mixins.toolbar,
-    justifyContent: "flex-end"
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing.unit * 3,
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen
-    }),
-    marginLeft: -drawerWidth
-  },
-  contentShift: {
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen
-    }),
-    marginLeft: 0
-  }
-});
+import styles from '../css/styles'
 
 class Header extends Component {
   state = {
@@ -103,7 +44,7 @@ class Header extends Component {
     return null
     }    
     return (
-      <div className={classes.root}>
+      <div className={classes.rootHead}>
         <CssBaseline />
         <AppBar
           position="fixed"
