+++ client/src/components/forms/onboard/startOnboard.js
@@ -10,7 +10,6 @@ import { connect } from "react-redux";
 import "react-datepicker/dist/react-datepicker.css";
 import * as Validation from "../../../constants/ValidationOptions";
 import Container from "@material-ui/core/Container";
-import submit from "../../../constants/SubmitValidation";
 import { Message } from 'semantic-ui-react'
