+++ client/src/css/styles.js
@@ -1,0 +1,69 @@
+const drawerWidth = 240;
+
+const styles = theme => ({
+    root: {
+      width: "90%",
+      margin: "0 auto"
+    },
+    button: {
+      margin: theme.spacing(1),
+    },
+    rootHead: {
+      display: "flex",
+      marginBottom: theme.spacing.unit * 15
+    },
+    appBar: {
+      transition: theme.transitions.create(["margin", "width"], {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen
+      })
+    },
+    appBarShift: {
+      width: `calc(100% - ${drawerWidth}px)`,
+      marginLeft: drawerWidth,
+      transition: theme.transitions.create(["margin", "width"], {
+        easing: theme.transitions.easing.easeOut,
+        duration: theme.transitions.duration.enteringScreen
+      })
+    },
+    menuButton: {
+      marginLeft: 12,
+      marginRight: 20
+    },
+    hide: {
+      display: "none"
+    },
+    drawer: {
+      width: drawerWidth,
+      flexShrink: 0
+    },
+    drawerPaper: {
+      width: drawerWidth
+    },
+    drawerHeader: {
+      display: "flex",
+      alignItems: "center",
+      padding: "0 8px",
+      ...theme.mixins.toolbar,
+      justifyContent: "flex-end"
+    },
+    content: {
+      flexGrow: 1,
+      padding: theme.spacing.unit * 3,
+      transition: theme.transitions.create("margin", {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen
+      }),
+      marginLeft: -drawerWidth
+    },
+    contentShift: {
+      transition: theme.transitions.create("margin", {
+        easing: theme.transitions.easing.easeOut,
+        duration: theme.transitions.duration.enteringScreen
+      }),
+      marginLeft: 0
+    }
+});
+
+export default styles;
+
